third_party/WebKit/LayoutTests/editing/selection/extend/extend_selection_character_20.html

<!doctype html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../assert_selection.js"></script><script>
function testIt(dir, offset, direction, expectedText) {
  test(() => assert_selection(
  `<span dir="${dir}">${makeSample(offset)}</span>`,
  selection => selection.modify('extend', direction, 'character'),
  `<span dir="${dir}">${expectedText}</span>`),
  `${dir} ${offset} extend ${direction} character`
  );
}

function makeSample(offset){
  const kSample = '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br />edf \u05D6;\u05D7;\u05D8; abrebg';
  return kSample.substr(0, offset) + '|' + kSample.substring(offset);
}

// ltr right character
testIt('ltr', 0, 'right', '^\u05D0|;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('ltr', 1, 'right', '\u05D0^;|\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('ltr', 2, 'right', '\u05D0;^\u05D1|;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('ltr', 3, 'right', '\u05D0;\u05D1^;|\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('ltr', 4, 'right', '\u05D0;\u05D1;^\u05D2|; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('ltr', 5, 'right', '\u05D0;\u05D1;\u05D2^;| abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('ltr', 6, 'right', '\u05D0;\u05D1;\u05D2;^ |abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('ltr', 7, 'right', '\u05D0;\u05D1;\u05D2; ^a|bc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('ltr', 8, 'right', '\u05D0;\u05D1;\u05D2; a^b|c \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('ltr', 9, 'right', '\u05D0;\u05D1;\u05D2; ab^c| \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('ltr', 10, 'right', '\u05D0;\u05D1;\u05D2; abc^ |\u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('ltr', 11, 'right', '\u05D0;\u05D1;\u05D2; abc ^\u05D3|;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('ltr', 12, 'right', '\u05D0;\u05D1;\u05D2; abc \u05D3^;|\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('ltr', 13, 'right', '\u05D0;\u05D1;\u05D2; abc \u05D3;^\u05D4|;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('ltr', 14, 'right', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4^;|\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('ltr', 15, 'right', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;^\u05D5|;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('ltr', 16, 'right', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5^;|<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('ltr', 17, 'right', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;^<br>|edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('ltr', 18, 'right', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<^b|r />edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('ltr', 19, 'right', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<b|r>edf \u05D6;\u05D7;\u05D8; abrebg</b|r>');
testIt('ltr', 20, 'right', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br|>edf \u05D6;\u05D7;\u05D8; abrebg</br|>');
testIt('ltr', 21, 'right', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br |>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('ltr', 22, 'right', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br |>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('ltr', 23, 'right', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>^e|df \u05D6;\u05D7;\u05D8; abrebg');
testIt('ltr', 24, 'right', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>e^d|f \u05D6;\u05D7;\u05D8; abrebg');
testIt('ltr', 25, 'right', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>ed^f| \u05D6;\u05D7;\u05D8; abrebg');
testIt('ltr', 26, 'right', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf^ |\u05D6;\u05D7;\u05D8; abrebg');
testIt('ltr', 27, 'right', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf ^\u05D6|;\u05D7;\u05D8; abrebg');
testIt('ltr', 28, 'right', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6^;|\u05D7;\u05D8; abrebg');
testIt('ltr', 29, 'right', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;^\u05D7|;\u05D8; abrebg');
testIt('ltr', 30, 'right', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7^;|\u05D8; abrebg');
testIt('ltr', 31, 'right', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;^\u05D8|; abrebg');
testIt('ltr', 32, 'right', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8^;| abrebg');
testIt('ltr', 33, 'right', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8;^ |abrebg');
testIt('ltr', 34, 'right', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; ^a|brebg');
testIt('ltr', 35, 'right', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; a^b|rebg');
testIt('ltr', 36, 'right', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; ab^r|ebg');
testIt('ltr', 37, 'right', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abr^e|bg');
testIt('ltr', 38, 'right', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abre^b|g');
testIt('ltr', 39, 'right', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abreb^g|');

// ltr left character
testIt('ltr', 0, 'left', '|\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('ltr', 1, 'left', '|\u05D0^;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('ltr', 2, 'left', '\u05D0|;^\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('ltr', 3, 'left', '\u05D0;|\u05D1^;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('ltr', 4, 'left', '\u05D0;\u05D1|;^\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('ltr', 5, 'left', '\u05D0;\u05D1;|\u05D2^; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('ltr', 6, 'left', '\u05D0;\u05D1;\u05D2|;^ abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('ltr', 7, 'left', '\u05D0;\u05D1;\u05D2;| ^abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('ltr', 8, 'left', '\u05D0;\u05D1;\u05D2; |a^bc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('ltr', 9, 'left', '\u05D0;\u05D1;\u05D2; a|b^c \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('ltr', 10, 'left', '\u05D0;\u05D1;\u05D2; ab|c^ \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('ltr', 11, 'left', '\u05D0;\u05D1;\u05D2; abc| ^\u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('ltr', 12, 'left', '\u05D0;\u05D1;\u05D2; abc |\u05D3^;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('ltr', 13, 'left', '\u05D0;\u05D1;\u05D2; abc \u05D3|;^\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('ltr', 14, 'left', '\u05D0;\u05D1;\u05D2; abc \u05D3;|\u05D4^;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('ltr', 15, 'left', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4|;^\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('ltr', 16, 'left', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;|\u05D5^;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('ltr', 17, 'left', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5|;^<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('ltr', 18, 'left', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;|<^br />edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('ltr', 19, 'left', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<b|r>edf \u05D6;\u05D7;\u05D8; abrebg</b|r>');
testIt('ltr', 20, 'left', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br|>edf \u05D6;\u05D7;\u05D8; abrebg</br|>');
testIt('ltr', 21, 'left', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br |>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('ltr', 22, 'left', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br |>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('ltr', 23, 'left', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;|<br>^edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('ltr', 24, 'left', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>|e^df \u05D6;\u05D7;\u05D8; abrebg');
testIt('ltr', 25, 'left', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>e|d^f \u05D6;\u05D7;\u05D8; abrebg');
testIt('ltr', 26, 'left', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>ed|f^ \u05D6;\u05D7;\u05D8; abrebg');
testIt('ltr', 27, 'left', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf| ^\u05D6;\u05D7;\u05D8; abrebg');
testIt('ltr', 28, 'left', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf |\u05D6^;\u05D7;\u05D8; abrebg');
testIt('ltr', 29, 'left', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6|;^\u05D7;\u05D8; abrebg');
testIt('ltr', 30, 'left', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;|\u05D7^;\u05D8; abrebg');
testIt('ltr', 31, 'left', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7|;^\u05D8; abrebg');
testIt('ltr', 32, 'left', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;|\u05D8^; abrebg');
testIt('ltr', 33, 'left', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8|;^ abrebg');
testIt('ltr', 34, 'left', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8;| ^abrebg');
testIt('ltr', 35, 'left', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; |a^brebg');
testIt('ltr', 36, 'left', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; a|b^rebg');
testIt('ltr', 37, 'left', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; ab|r^ebg');
testIt('ltr', 38, 'left', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abr|e^bg');
testIt('ltr', 39, 'left', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abre|b^g');

// ltr forward character
testIt('ltr', 0, 'forward', '^\u05D0|;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('ltr', 1, 'forward', '\u05D0^;|\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('ltr', 2, 'forward', '\u05D0;^\u05D1|;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('ltr', 3, 'forward', '\u05D0;\u05D1^;|\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('ltr', 4, 'forward', '\u05D0;\u05D1;^\u05D2|; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('ltr', 5, 'forward', '\u05D0;\u05D1;\u05D2^;| abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('ltr', 6, 'forward', '\u05D0;\u05D1;\u05D2;^ |abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('ltr', 7, 'forward', '\u05D0;\u05D1;\u05D2; ^a|bc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('ltr', 8, 'forward', '\u05D0;\u05D1;\u05D2; a^b|c \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('ltr', 9, 'forward', '\u05D0;\u05D1;\u05D2; ab^c| \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('ltr', 10, 'forward', '\u05D0;\u05D1;\u05D2; abc^ |\u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('ltr', 11, 'forward', '\u05D0;\u05D1;\u05D2; abc ^\u05D3|;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('ltr', 12, 'forward', '\u05D0;\u05D1;\u05D2; abc \u05D3^;|\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('ltr', 13, 'forward', '\u05D0;\u05D1;\u05D2; abc \u05D3;^\u05D4|;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('ltr', 14, 'forward', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4^;|\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('ltr', 15, 'forward', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;^\u05D5|;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('ltr', 16, 'forward', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5^;|<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('ltr', 17, 'forward', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;^<br>|edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('ltr', 18, 'forward', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<^b|r />edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('ltr', 19, 'forward', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<b|r>edf \u05D6;\u05D7;\u05D8; abrebg</b|r>');
testIt('ltr', 20, 'forward', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br|>edf \u05D6;\u05D7;\u05D8; abrebg</br|>');
testIt('ltr', 21, 'forward', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br |>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('ltr', 22, 'forward', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br |>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('ltr', 23, 'forward', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>^e|df \u05D6;\u05D7;\u05D8; abrebg');
testIt('ltr', 24, 'forward', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>e^d|f \u05D6;\u05D7;\u05D8; abrebg');
testIt('ltr', 25, 'forward', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>ed^f| \u05D6;\u05D7;\u05D8; abrebg');
testIt('ltr', 26, 'forward', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf^ |\u05D6;\u05D7;\u05D8; abrebg');
testIt('ltr', 27, 'forward', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf ^\u05D6|;\u05D7;\u05D8; abrebg');
testIt('ltr', 28, 'forward', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6^;|\u05D7;\u05D8; abrebg');
testIt('ltr', 29, 'forward', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;^\u05D7|;\u05D8; abrebg');
testIt('ltr', 30, 'forward', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7^;|\u05D8; abrebg');
testIt('ltr', 31, 'forward', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;^\u05D8|; abrebg');
testIt('ltr', 32, 'forward', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8^;| abrebg');
testIt('ltr', 33, 'forward', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8;^ |abrebg');
testIt('ltr', 34, 'forward', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; ^a|brebg');
testIt('ltr', 35, 'forward', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; a^b|rebg');
testIt('ltr', 36, 'forward', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; ab^r|ebg');
testIt('ltr', 37, 'forward', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abr^e|bg');
testIt('ltr', 38, 'forward', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abre^b|g');
testIt('ltr', 39, 'forward', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abreb^g|');

// ltr backward character
testIt('ltr', 0, 'backward', '|\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('ltr', 1, 'backward', '|\u05D0^;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('ltr', 2, 'backward', '\u05D0|;^\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('ltr', 3, 'backward', '\u05D0;|\u05D1^;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('ltr', 4, 'backward', '\u05D0;\u05D1|;^\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('ltr', 5, 'backward', '\u05D0;\u05D1;|\u05D2^; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('ltr', 6, 'backward', '\u05D0;\u05D1;\u05D2|;^ abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('ltr', 7, 'backward', '\u05D0;\u05D1;\u05D2;| ^abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('ltr', 8, 'backward', '\u05D0;\u05D1;\u05D2; |a^bc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('ltr', 9, 'backward', '\u05D0;\u05D1;\u05D2; a|b^c \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('ltr', 10, 'backward', '\u05D0;\u05D1;\u05D2; ab|c^ \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('ltr', 11, 'backward', '\u05D0;\u05D1;\u05D2; abc| ^\u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('ltr', 12, 'backward', '\u05D0;\u05D1;\u05D2; abc |\u05D3^;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('ltr', 13, 'backward', '\u05D0;\u05D1;\u05D2; abc \u05D3|;^\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('ltr', 14, 'backward', '\u05D0;\u05D1;\u05D2; abc \u05D3;|\u05D4^;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('ltr', 15, 'backward', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4|;^\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('ltr', 16, 'backward', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;|\u05D5^;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('ltr', 17, 'backward', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5|;^<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('ltr', 18, 'backward', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;|<^br />edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('ltr', 19, 'backward', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<b|r>edf \u05D6;\u05D7;\u05D8; abrebg</b|r>');
testIt('ltr', 20, 'backward', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br|>edf \u05D6;\u05D7;\u05D8; abrebg</br|>');
testIt('ltr', 21, 'backward', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br |>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('ltr', 22, 'backward', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br |>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('ltr', 23, 'backward', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;|<br>^edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('ltr', 24, 'backward', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>|e^df \u05D6;\u05D7;\u05D8; abrebg');
testIt('ltr', 25, 'backward', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>e|d^f \u05D6;\u05D7;\u05D8; abrebg');
testIt('ltr', 26, 'backward', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>ed|f^ \u05D6;\u05D7;\u05D8; abrebg');
testIt('ltr', 27, 'backward', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf| ^\u05D6;\u05D7;\u05D8; abrebg');
testIt('ltr', 28, 'backward', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf |\u05D6^;\u05D7;\u05D8; abrebg');
testIt('ltr', 29, 'backward', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6|;^\u05D7;\u05D8; abrebg');
testIt('ltr', 30, 'backward', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;|\u05D7^;\u05D8; abrebg');
testIt('ltr', 31, 'backward', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7|;^\u05D8; abrebg');
testIt('ltr', 32, 'backward', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;|\u05D8^; abrebg');
testIt('ltr', 33, 'backward', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8|;^ abrebg');
testIt('ltr', 34, 'backward', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8;| ^abrebg');
testIt('ltr', 35, 'backward', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; |a^brebg');
testIt('ltr', 36, 'backward', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; a|b^rebg');
testIt('ltr', 37, 'backward', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; ab|r^ebg');
testIt('ltr', 38, 'backward', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abr|e^bg');
testIt('ltr', 39, 'backward', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abre|b^g');

// rtl right character
testIt('rtl', 0, 'right', '^\u05D0|;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('rtl', 1, 'right', '\u05D0^;|\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('rtl', 2, 'right', '\u05D0;^\u05D1|;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('rtl', 3, 'right', '\u05D0;\u05D1^;|\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('rtl', 4, 'right', '\u05D0;\u05D1;^\u05D2|; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('rtl', 5, 'right', '\u05D0;\u05D1;\u05D2^;| abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('rtl', 6, 'right', '\u05D0;\u05D1;\u05D2;^ |abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('rtl', 7, 'right', '\u05D0;\u05D1;\u05D2; ^a|bc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('rtl', 8, 'right', '\u05D0;\u05D1;\u05D2; a^b|c \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('rtl', 9, 'right', '\u05D0;\u05D1;\u05D2; ab^c| \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('rtl', 10, 'right', '\u05D0;\u05D1;\u05D2; abc^ |\u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('rtl', 11, 'right', '\u05D0;\u05D1;\u05D2; abc ^\u05D3|;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('rtl', 12, 'right', '\u05D0;\u05D1;\u05D2; abc \u05D3^;|\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('rtl', 13, 'right', '\u05D0;\u05D1;\u05D2; abc \u05D3;^\u05D4|;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('rtl', 14, 'right', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4^;|\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('rtl', 15, 'right', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;^\u05D5|;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('rtl', 16, 'right', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5^;|<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('rtl', 17, 'right', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;^<br>|edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('rtl', 18, 'right', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<^b|r />edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('rtl', 19, 'right', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<b|r>edf \u05D6;\u05D7;\u05D8; abrebg</b|r>');
testIt('rtl', 20, 'right', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br|>edf \u05D6;\u05D7;\u05D8; abrebg</br|>');
testIt('rtl', 21, 'right', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br |>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('rtl', 22, 'right', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br |>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('rtl', 23, 'right', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>^e|df \u05D6;\u05D7;\u05D8; abrebg');
testIt('rtl', 24, 'right', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>e^d|f \u05D6;\u05D7;\u05D8; abrebg');
testIt('rtl', 25, 'right', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>ed^f| \u05D6;\u05D7;\u05D8; abrebg');
testIt('rtl', 26, 'right', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf^ |\u05D6;\u05D7;\u05D8; abrebg');
testIt('rtl', 27, 'right', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf ^\u05D6|;\u05D7;\u05D8; abrebg');
testIt('rtl', 28, 'right', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6^;|\u05D7;\u05D8; abrebg');
testIt('rtl', 29, 'right', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;^\u05D7|;\u05D8; abrebg');
testIt('rtl', 30, 'right', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7^;|\u05D8; abrebg');
testIt('rtl', 31, 'right', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;^\u05D8|; abrebg');
testIt('rtl', 32, 'right', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8^;| abrebg');
testIt('rtl', 33, 'right', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8;^ |abrebg');
testIt('rtl', 34, 'right', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; ^a|brebg');
testIt('rtl', 35, 'right', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; a^b|rebg');
testIt('rtl', 36, 'right', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; ab^r|ebg');
testIt('rtl', 37, 'right', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abr^e|bg');
testIt('rtl', 38, 'right', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abre^b|g');
testIt('rtl', 39, 'right', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abreb^g|');

// rtl left character
testIt('rtl', 0, 'left', '|\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('rtl', 1, 'left', '|\u05D0^;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('rtl', 2, 'left', '\u05D0|;^\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('rtl', 3, 'left', '\u05D0;|\u05D1^;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('rtl', 4, 'left', '\u05D0;\u05D1|;^\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('rtl', 5, 'left', '\u05D0;\u05D1;|\u05D2^; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('rtl', 6, 'left', '\u05D0;\u05D1;\u05D2|;^ abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('rtl', 7, 'left', '\u05D0;\u05D1;\u05D2;| ^abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('rtl', 8, 'left', '\u05D0;\u05D1;\u05D2; |a^bc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('rtl', 9, 'left', '\u05D0;\u05D1;\u05D2; a|b^c \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('rtl', 10, 'left', '\u05D0;\u05D1;\u05D2; ab|c^ \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('rtl', 11, 'left', '\u05D0;\u05D1;\u05D2; abc| ^\u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('rtl', 12, 'left', '\u05D0;\u05D1;\u05D2; abc |\u05D3^;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('rtl', 13, 'left', '\u05D0;\u05D1;\u05D2; abc \u05D3|;^\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('rtl', 14, 'left', '\u05D0;\u05D1;\u05D2; abc \u05D3;|\u05D4^;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('rtl', 15, 'left', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4|;^\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('rtl', 16, 'left', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;|\u05D5^;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('rtl', 17, 'left', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5|;^<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('rtl', 18, 'left', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;|<^br />edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('rtl', 19, 'left', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<b|r>edf \u05D6;\u05D7;\u05D8; abrebg</b|r>');
testIt('rtl', 20, 'left', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br|>edf \u05D6;\u05D7;\u05D8; abrebg</br|>');
testIt('rtl', 21, 'left', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br |>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('rtl', 22, 'left', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br |>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('rtl', 23, 'left', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;|<br>^edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('rtl', 24, 'left', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>|e^df \u05D6;\u05D7;\u05D8; abrebg');
testIt('rtl', 25, 'left', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>e|d^f \u05D6;\u05D7;\u05D8; abrebg');
testIt('rtl', 26, 'left', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>ed|f^ \u05D6;\u05D7;\u05D8; abrebg');
testIt('rtl', 27, 'left', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf| ^\u05D6;\u05D7;\u05D8; abrebg');
testIt('rtl', 28, 'left', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf |\u05D6^;\u05D7;\u05D8; abrebg');
testIt('rtl', 29, 'left', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6|;^\u05D7;\u05D8; abrebg');
testIt('rtl', 30, 'left', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;|\u05D7^;\u05D8; abrebg');
testIt('rtl', 31, 'left', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7|;^\u05D8; abrebg');
testIt('rtl', 32, 'left', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;|\u05D8^; abrebg');
testIt('rtl', 33, 'left', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8|;^ abrebg');
testIt('rtl', 34, 'left', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8;| ^abrebg');
testIt('rtl', 35, 'left', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; |a^brebg');
testIt('rtl', 36, 'left', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; a|b^rebg');
testIt('rtl', 37, 'left', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; ab|r^ebg');
testIt('rtl', 38, 'left', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abr|e^bg');
testIt('rtl', 39, 'left', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abre|b^g');

// rtl forward character
testIt('rtl', 0, 'forward', '^\u05D0|;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('rtl', 1, 'forward', '\u05D0^;|\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('rtl', 2, 'forward', '\u05D0;^\u05D1|;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('rtl', 3, 'forward', '\u05D0;\u05D1^;|\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('rtl', 4, 'forward', '\u05D0;\u05D1;^\u05D2|; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('rtl', 5, 'forward', '\u05D0;\u05D1;\u05D2^;| abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('rtl', 6, 'forward', '\u05D0;\u05D1;\u05D2;^ |abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('rtl', 7, 'forward', '\u05D0;\u05D1;\u05D2; ^a|bc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('rtl', 8, 'forward', '\u05D0;\u05D1;\u05D2; a^b|c \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('rtl', 9, 'forward', '\u05D0;\u05D1;\u05D2; ab^c| \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('rtl', 10, 'forward', '\u05D0;\u05D1;\u05D2; abc^ |\u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('rtl', 11, 'forward', '\u05D0;\u05D1;\u05D2; abc ^\u05D3|;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('rtl', 12, 'forward', '\u05D0;\u05D1;\u05D2; abc \u05D3^;|\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('rtl', 13, 'forward', '\u05D0;\u05D1;\u05D2; abc \u05D3;^\u05D4|;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('rtl', 14, 'forward', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4^;|\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('rtl', 15, 'forward', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;^\u05D5|;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('rtl', 16, 'forward', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5^;|<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('rtl', 17, 'forward', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;^<br>|edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('rtl', 18, 'forward', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<^b|r />edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('rtl', 19, 'forward', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<b|r>edf \u05D6;\u05D7;\u05D8; abrebg</b|r>');
testIt('rtl', 20, 'forward', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br|>edf \u05D6;\u05D7;\u05D8; abrebg</br|>');
testIt('rtl', 21, 'forward', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br |>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('rtl', 22, 'forward', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br |>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('rtl', 23, 'forward', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>^e|df \u05D6;\u05D7;\u05D8; abrebg');
testIt('rtl', 24, 'forward', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>e^d|f \u05D6;\u05D7;\u05D8; abrebg');
testIt('rtl', 25, 'forward', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>ed^f| \u05D6;\u05D7;\u05D8; abrebg');
testIt('rtl', 26, 'forward', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf^ |\u05D6;\u05D7;\u05D8; abrebg');
testIt('rtl', 27, 'forward', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf ^\u05D6|;\u05D7;\u05D8; abrebg');
testIt('rtl', 28, 'forward', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6^;|\u05D7;\u05D8; abrebg');
testIt('rtl', 29, 'forward', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;^\u05D7|;\u05D8; abrebg');
testIt('rtl', 30, 'forward', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7^;|\u05D8; abrebg');
testIt('rtl', 31, 'forward', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;^\u05D8|; abrebg');
testIt('rtl', 32, 'forward', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8^;| abrebg');
testIt('rtl', 33, 'forward', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8;^ |abrebg');
testIt('rtl', 34, 'forward', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; ^a|brebg');
testIt('rtl', 35, 'forward', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; a^b|rebg');
testIt('rtl', 36, 'forward', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; ab^r|ebg');
testIt('rtl', 37, 'forward', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abr^e|bg');
testIt('rtl', 38, 'forward', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abre^b|g');
testIt('rtl', 39, 'forward', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abreb^g|');

// rtl backward character
testIt('rtl', 0, 'backward', '|\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('rtl', 1, 'backward', '|\u05D0^;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('rtl', 2, 'backward', '\u05D0|;^\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('rtl', 3, 'backward', '\u05D0;|\u05D1^;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('rtl', 4, 'backward', '\u05D0;\u05D1|;^\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('rtl', 5, 'backward', '\u05D0;\u05D1;|\u05D2^; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('rtl', 6, 'backward', '\u05D0;\u05D1;\u05D2|;^ abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('rtl', 7, 'backward', '\u05D0;\u05D1;\u05D2;| ^abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('rtl', 8, 'backward', '\u05D0;\u05D1;\u05D2; |a^bc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('rtl', 9, 'backward', '\u05D0;\u05D1;\u05D2; a|b^c \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('rtl', 10, 'backward', '\u05D0;\u05D1;\u05D2; ab|c^ \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('rtl', 11, 'backward', '\u05D0;\u05D1;\u05D2; abc| ^\u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('rtl', 12, 'backward', '\u05D0;\u05D1;\u05D2; abc |\u05D3^;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('rtl', 13, 'backward', '\u05D0;\u05D1;\u05D2; abc \u05D3|;^\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('rtl', 14, 'backward', '\u05D0;\u05D1;\u05D2; abc \u05D3;|\u05D4^;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('rtl', 15, 'backward', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4|;^\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('rtl', 16, 'backward', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;|\u05D5^;<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('rtl', 17, 'backward', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5|;^<br>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('rtl', 18, 'backward', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;|<^br />edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('rtl', 19, 'backward', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<b|r>edf \u05D6;\u05D7;\u05D8; abrebg</b|r>');
testIt('rtl', 20, 'backward', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br|>edf \u05D6;\u05D7;\u05D8; abrebg</br|>');
testIt('rtl', 21, 'backward', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br |>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('rtl', 22, 'backward', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br |>edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('rtl', 23, 'backward', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;|<br>^edf \u05D6;\u05D7;\u05D8; abrebg');
testIt('rtl', 24, 'backward', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>|e^df \u05D6;\u05D7;\u05D8; abrebg');
testIt('rtl', 25, 'backward', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>e|d^f \u05D6;\u05D7;\u05D8; abrebg');
testIt('rtl', 26, 'backward', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>ed|f^ \u05D6;\u05D7;\u05D8; abrebg');
testIt('rtl', 27, 'backward', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf| ^\u05D6;\u05D7;\u05D8; abrebg');
testIt('rtl', 28, 'backward', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf |\u05D6^;\u05D7;\u05D8; abrebg');
testIt('rtl', 29, 'backward', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6|;^\u05D7;\u05D8; abrebg');
testIt('rtl', 30, 'backward', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;|\u05D7^;\u05D8; abrebg');
testIt('rtl', 31, 'backward', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7|;^\u05D8; abrebg');
testIt('rtl', 32, 'backward', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;|\u05D8^; abrebg');
testIt('rtl', 33, 'backward', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8|;^ abrebg');
testIt('rtl', 34, 'backward', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8;| ^abrebg');
testIt('rtl', 35, 'backward', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; |a^brebg');
testIt('rtl', 36, 'backward', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; a|b^rebg');
testIt('rtl', 37, 'backward', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; ab|r^ebg');
testIt('rtl', 38, 'backward', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abr|e^bg');
testIt('rtl', 39, 'backward', '\u05D0;\u05D1;\u05D2; abc \u05D3;\u05D4;\u05D5;<br>edf \u05D6;\u05D7;\u05D8; abre|b^g');
</script>
